refactor(lsp): split call hierarchy item builder
